Nothing interesting here, just getting some things ready before I refactor Promises in Unsplash API
