--- app/ext-api/handleAPIS.js
@@ -1,4 +1,0 @@
-// const unsplashPic = require('./app/ext-api/unsplash/unsplash.js');
-//
-//
-// console.log(unsplashPic);
